6/6/24 code pushed

+++ app/Http/Controllers/User/PostController.php
@@ -144,14 +144,14 @@ class PostController extends Controller
                 [
                     'title' => 'required',
                     'description' => 'nullable|required_without:media',
-                    'media' => 'nullable',
+                    'media' => 'nullable|required_if:on_instagram,1',
                     'on_facebook' => 'nullable|boolean',
                     'facebook_page' => 'nullable|required_if:on_facebook,1',
                     'on_instagram' => 'nullable|boolean',
                     'instagram_account' => 'nullable|required_if:on_instagram,1',
                     'on_linkedin' => 'nullable|boolean',
                     'linkedin_organization' => 'nullable|required_if:on_linkedin,1',
-                    'schedule_date' => 'nullable|date|after:today',
+                    'schedule_date' => 'nullable|date|after_or_equal:today',
                     'schedule_time' => 'nullable|date_format:H:i',
                 ],
                 [
@@ -161,7 +161,7 @@ class PostController extends Controller
                     'description.required_without' => 'Description is required',
 
                     'media.required' => 'Media is required',
-                    'media.required_if' => 'Media is required',
+                    'media.required_if' => 'Media is required for Instagram.',
 
                     'facebook_page.required_if' => 'Facebook Page is required.',
 
@@ -169,9 +169,9 @@ class PostController extends Controller
 
                     'linkedin_organization.required_if' => 'Linkedin Organization is required.',
 
-                    'schedule_date.after' => 'Schedule date should be a future date.',
+                    'schedule_date.after_or_equal' => 'Schedule date should be a future date.',
 
-                    'schedule_time.after' => 'Schedule time should be a future time.',
+                    'schedule_time.after_or_equal' => 'Schedule time should be a future time.',
                 ]
             );
 
@@ -210,9 +210,11 @@ class PostController extends Controller
                 $facebook_page_access_token = $facebook_page[1];
                 $facebook_page_name = $facebook_page[2];
 
+                $long_facebook_access_token = $this->facebookService->tokenTime($facebook_page_access_token);
+
                 $data->on_facebook = 1;
                 $data->facebook_page_id = $facebook_page_id;
-                $data->facebook_page_access_token = $facebook_page_access_token;
+                $data->facebook_page_access_token = $long_facebook_access_token;
                 $data->facebook_page_name = $facebook_page_name;
             }
 
@@ -243,6 +245,14 @@ class PostController extends Controller
             } else {
                 $assets = env('APP_URL') . '/';
 
+                // On Instagram
+                if ($request->has('on_instagram')) {
+                    if ($request->hasFile('media')) {
+                        // if ($media_type == 'image') $this->instagramService->postImage($data->instagram_account_id, $assets . $data->media, $request->description);
+                        // if ($media_type == 'video') $this->instagramService->postVideo($data->instagram_account_id, $assets . $data->media, $mediaSize, $request->description);
+                    }
+                }
+
                 // On Facebook
                 if ($request->has('on_facebook')) {
                     if ($request->hasFile('media')) {
@@ -259,15 +269,7 @@ class PostController extends Controller
                         // if ($media_type == 'image') $this->linkedinService->postImage($data->linkedin_company_id, $assets . $data->media, $request->description);
                         // if ($media_type == 'video') $this->linkedinService->postVideo($data->linkedin_company_id, $assets . $data->media, $request->description);
                     } else {
-                        $this->linkedinService->postText($data->linkedin_company_id, $request->description);
-                    }
-                }
-
-                // On Instagram
-                if ($request->has('on_instagram')) {
-                    if ($request->hasFile('media')) {
-                        // if ($media_type == 'image') $this->instagramService->postImage($data->instagram_account_id, $assets . $data->media, $request->description);
-                        // if ($media_type == 'video') $this->instagramService->postVideo($data->instagram_account_id, $assets . $data->media, $mediaSize, $request->description);
+                        // $this->linkedinService->postText($data->linkedin_company_id, $request->description);
                     }
                 }
 
@@ -278,9 +280,14 @@ class PostController extends Controller
 
             DB::commit();
 
+            if ($request->schedule_date != null && $request->schedule_time != null) return response()->json([
+                'status' => 200,
+                'message' => 'Post scheduled successfully'
+            ]);
+
             return response()->json([
                 'status' => 200,
-                'message' => 'Posted successfully'
+                'message' => 'Post published successfully'
             ]);
         } catch (Exception $e) {
             DB::rollBack();
@@ -305,23 +312,38 @@ class PostController extends Controller
                 [
                     'post_id' => 'required|exists:posts,id',
                     'title' => 'required',
-                    'description' => 'nullable',
+                    'description' => 'nullable|required_without:media',
                     'media' => 'nullable',
                     'on_facebook' => 'nullable|boolean',
+                    'facebook_page' => 'nullable|required_if:on_facebook,1',
                     'on_instagram' => 'nullable|boolean',
+                    'instagram_account' => 'nullable|required_if:on_instagram,1',
                     'on_linkedin' => 'nullable|boolean',
-                    'schedule_date' => 'nullable|date|after:today',
+                    'linkedin_organization' => 'nullable|required_if:on_linkedin,1',
+                    'schedule_date' => 'nullable|date|after_or_equal:today',
                     'schedule_time' => 'nullable|date_format:H:i',
                 ],
                 [
+                    'post_id.required' => 'Invalid Request',
+                    'post_id.exists' => 'Invalid Request',
+
                     'title.required' => 'Title is required',
 
                     'description.required' => 'Description is required',
                     'description.required_without' => 'Description is required',
 
-                    'schedule_date.after' => 'Schedule date should be a future date.',
+                    'media.required' => 'Media is required',
+                    'media.required_if' => 'Media is required',
 
-                    'schedule_time.after' => 'Schedule time should be a future time.',
+                    'facebook_page.required_if' => 'Facebook Page is required.',
+
+                    'instagram_account.required_if' => 'Instagram Account is required.',
+
+                    'linkedin_organization.required_if' => 'Linkedin Organization is required.',
+
+                    'schedule_date.after_or_equal' => 'Schedule date should be a future date.',
+
+                    'schedule_time.after_or_equal' => 'Schedule time should be a future time.',
                 ]
             );
 
@@ -332,9 +354,8 @@ class PostController extends Controller
             $data = new Post;
             $data->user_id = Auth::guard('web')->user()->id;
             $data->title = $request->title;
-            $data->description = $request->description;
+            $data->description = str_replace('\n', "\n", $request->description);
 
-            $mediaPath = null;
             $mediaType = null;
             $media_type = null;
 
@@ -361,7 +382,6 @@ class PostController extends Controller
                     copy(public_path($oldPost->media), public_path('posts') . '/' . $mediaName);
 
                     $mediaSize = filesize(public_path($onlyMediaPath));
-                    $mediaPath = public_path('posts') . '/' . $mediaName;
 
                     $media_type = $oldPost->media_type;
                     $data->media = $onlyMediaPath;
@@ -369,35 +389,73 @@ class PostController extends Controller
                 }
             }
 
-            if ($request->has('on_facebook')) $data->on_facebook = 1;
-            if ($request->has('on_instagram')) $data->on_instagram = 1;
-            if ($request->has('on_linkedin')) $data->on_linkedin = 1;
+            // On Facebook
+            if ($request->has('on_facebook')) {
+                $facebook_page = explode(' - ', $request->facebook_page);
+                $facebook_page_id = $facebook_page[0];
+                $facebook_page_access_token = $facebook_page[1];
+                $facebook_page_name = $facebook_page[2];
+
+                $long_facebook_access_token = $this->facebookService->tokenTime($facebook_page_access_token);
+
+                $data->on_facebook = 1;
+                $data->facebook_page_id = $facebook_page_id;
+                $data->facebook_page_access_token = $long_facebook_access_token;
+                $data->facebook_page_name = $facebook_page_name;
+            }
+
+            // On Instagram
+            if ($request->has('on_instagram')) {
+                $instagram_account = explode(' - ', $request->instagram_account);
+                $instagram_account_id = $instagram_account[0];
+                $instagram_account_name = $instagram_account[1];
+
+                $data->on_instagram = 1;
+                $data->instagram_account_id = $instagram_account_id;
+                $data->instagram_account_name = $instagram_account_name;
+            }
+
+            // On Linkedin
+            if ($request->has('on_linkedin')) {
+                $linkedin_organization = explode(' - ', $request->linkedin_organization);
+                $linkedin_organization_id = $linkedin_organization[0];
+                $linkedin_organization_name = $linkedin_organization[1];
+
+                $data->on_linkedin = 1;
+                $data->linkedin_company_id = $linkedin_organization_id;
+                $data->linkedin_company_name = $linkedin_organization_name;
+            }
 
             if ($request->schedule_date != null && $request->schedule_time != null) {
                 $data->scheduled_at = $request->schedule_date . ' ' . $request->schedule_time;
             } else {
-                if ($request->has('on_facebook')) {
-                    $page = explode(' - ', $request->facebook_page);
-                    $page_id = $page[0];
-                    $page_access_token = $page[1];
+                $assets = env('APP_URL') . '/';
 
+                // On Instagram
+                if ($request->has('on_instagram')) {
                     if ($request->hasFile('media')) {
-                        if ($media_type == 'image') {
-                            $post = $this->facebookService->postImage($page_id, $page_access_token, env('APP_URL') . '/' . $onlyMediaPath, $request->description);
-                        } elseif ($media_type == 'video') {
-                            $post = $this->facebookService->postVideo($page_id, $page_access_token, $mediaSize, env('APP_URL') . '/' . $onlyMediaPath, $request->description);
-                        }
-                    } else {
-                        $post = $this->facebookService->postText($page_id, $page_access_token, $request->description);
+                        // if ($media_type == 'image') $this->instagramService->postImage($data->instagram_account_id, $assets . $data->media, $request->description);
+                        // if ($media_type == 'video') $this->instagramService->postVideo($data->instagram_account_id, $assets . $data->media, $mediaSize, $request->description);
                     }
                 }
 
+                // On Facebook
+                if ($request->has('on_facebook')) {
+                    if ($request->hasFile('media')) {
+                        // if ($media_type == 'image') $this->facebookService->postImage($data->facebook_page_id, $data->facebook_page_access_token, $assets . $data->media, $request->description);
+                        // if ($media_type == 'video') $this->facebookService->postVideo($data->facebook_page_id, $data->facebook_page_access_token, $mediaSize, $assets . $data->media, $request->description);
+                    } else {
+                        // $this->facebookService->postText($data->facebook_page_id, $data->facebook_page_access_token, $request->description);
+                    }
+                }
+
+                // On Linkedin
                 if ($request->has('on_linkedin')) {
                     if ($request->hasFile('media')) {
-                        if ($media_type == 'image') $post = $this->linkedinService->postImage($request->linkedin_organization, env('APP_URL') . '/' . $onlyMediaPath, $request->description);
-                        if ($media_type == 'video') $post = $this->linkedinService->postVideo($request->linkedin_organization, env('APP_URL') . '/' . $onlyMediaPath, $request->description);
+                        // if ($media_type == 'image') $this->linkedinService->postImage($data->linkedin_company_id, $assets . $data->media, $request->description);
+                        // if ($media_type == 'video') $this->linkedinService->postVideo($data->linkedin_company_id, $assets . $data->media, $request->description);
                     } else {
-                        $post = $this->linkedinService->postText($request->linkedin_organization, $request->description);
+                        // $this->linkedinService->postText($data->linkedin_company_id, $request->description);
                     }
                 }
 
@@ -408,7 +466,7 @@ class PostController extends Controller
 
             $oldPost = Post::where('id', $request->post_id)->first();
             if ($oldPost->draft == 1) {
-                unlink(public_path($oldPost->media));
+                if ($oldPost->media != null) if (file_exists(public_path($oldPost->media))) unlink(public_path($oldPost->media));
                 $oldPost->delete();
             }
 
@@ -447,8 +505,9 @@ class PostController extends Controller
             if ($data->media != null) if (file_exists(public_path($data->media))) unlink(public_path($data->media));
             $data->delete();
 
-            Session::flash('success', 'Post deleted successfully');
-            return redirect()->back();
+            return response()->json([
+                'status' => 200,
+            ]);
         } catch (Exception $e) {
             return response()->json([
                 'status' => 500,
